Harden keybinding implementation structure
