+++ src/app/app-shell/keybinding-events.ts
@@ -1,3 +1,4 @@
+import { useEffect, useRef } from 'react'
 import type { KeybindingCommandId } from '../../../shared/keybindings'
 
 export const howcodeKeybindingCommandEvent = 'howcode:keybinding-command'
@@ -13,4 +14,23 @@ export function dispatchHowcodeKeybindingCommand(commandId: KeybindingCommandId)
       detail: { commandId },
     }),
   )
+}
+
+export function useHowcodeKeybindingCommand(
+  commandId: KeybindingCommandId,
+  handler: (event: CustomEvent<HowcodeKeybindingCommandDetail>) => void,
+) {
+  const handlerRef = useRef(handler)
+  handlerRef.current = handler
+
+  useEffect(() => {
+    const handleCommand = (event: Event) => {
+      const customEvent = event as CustomEvent<HowcodeKeybindingCommandDetail>
+      if (customEvent.detail?.commandId !== commandId) return
+      handlerRef.current(customEvent)
+    }
+
+    window.addEventListener(howcodeKeybindingCommandEvent, handleCommand)
+    return () => window.removeEventListener(howcodeKeybindingCommandEvent, handleCommand)
+  }, [commandId])
 }
