+++ src/app/components/sidebar/chat/sidebar-chat-section.tsx
@@ -2,9 +2,8 @@ import { closestCorners, DndContext } from '@dnd-kit/core'
 import { restrictToParentElement, restrictToVerticalAxis } from '@dnd-kit/modifiers'
 import { SortableContext, verticalListSortingStrategy } from '@dnd-kit/sortable'
 import { Edit2, FolderPlus, Plus, Search } from 'lucide-react'
-import { useEffect, useRef } from 'react'
-import type { HowcodeKeybindingCommandDetail } from '../../../app-shell/keybinding-events'
-import { howcodeKeybindingCommandEvent } from '../../../app-shell/keybinding-events'
+import { useRef } from 'react'
+import { useHowcodeKeybindingCommand } from '../../../app-shell/keybinding-events'
 import type {
   ChatGroup,
   ChatSidebarState,
@@ -70,18 +69,11 @@ export function SidebarChatSection({
     ungroupedThreads,
   } = useChatSidebarController({ chatState, onCreateGroup, onRefresh, onAction })
 
-  useEffect(() => {
-    const handleCommand = (event: Event) => {
-      const commandId = (event as CustomEvent<HowcodeKeybindingCommandDetail>).detail?.commandId
-      if (commandId !== 'sidebar.find') return
-      event.preventDefault()
-      searchInputRef.current?.focus()
-      searchInputRef.current?.select()
-    }
-
-    window.addEventListener(howcodeKeybindingCommandEvent, handleCommand)
-    return () => window.removeEventListener(howcodeKeybindingCommandEvent, handleCommand)
-  }, [])
+  useHowcodeKeybindingCommand('sidebar.find', (event) => {
+    event.preventDefault()
+    searchInputRef.current?.focus()
+    searchInputRef.current?.select()
+  })
 
   const renderThread = (thread: ChatThread, groupId: string | null) => (
     <ChatThreadDropItem
